SLING-7447 - Race condition in ThreadLocalCleaner initialization code
Ensure that the reflection-obtained class fields are initialized eagerly
when the class is loaded. Since the class will be accessed only when the
thread pool is instantiated, the performance impact is minimal.
Moreover, this saves a synchronized check for each Runnable execution.
The gains are minimal, but there is less mutable state to reason about.